starting slide CRUD functions
